Finished GeneratingActivity.
Imitates the generation of a maze using a background thread.
While this is happening, the user can select a driver and robot configuration. The app will tell the user to wait for the generation if they select a driver before the thread is completed, and it will also tell the user to select a driver if the thread is completed and they haven't yet.